Fix: RuboCop fixes and add comments for Recipes::Fetcher
